requirements.txt
langchain
langchain-openai
langchain-community
chromadb
pypdf
streamlit
python-dotenv
